move StfStatePayload to Stf
