+++ stf/Cargo.toml
@@ -22,12 +22,13 @@ std = [
     "log",
     "base58",
     "sc-keystore",
+    "serde",
     "system/std",
     "sp-core/std",
     "hex",
     "substrate-api-client",
     "substrate-client-keystore",
-    "my-node-runtime"
+    "my-node-runtime",
 ]
 
 [dependencies]
@@ -39,6 +40,7 @@ log                 = { version = "0.4", optional = true }
 base58 			    = { version = "0.1", optional = true }
 derive_more         = { version = "0.99.5", optional = true }
 hex                 = { version = "0.4.2", optional = true }
+serde               = { version = "1.0", features = ["derive"], optional = true }
 
 [dependencies.sgx_tstd]
 git = "https://github.com/apache/teaclave-sgx-sdk.git"
