+++ stf/src/lib.rs
@@ -28,6 +28,9 @@ extern crate alloc;
 #[cfg(feature = "std")]
 extern crate clap;
 
+#[cfg(feature = "std")]
+use serde::{Deserialize, Serialize};
+
 use codec::{Compact, Decode, Encode};
 #[cfg(feature = "std")]
 use my_node_runtime::Balance;
@@ -275,6 +278,45 @@ pub struct TrustedReturnValue<T> {
 impl TrustedReturnValue
 */
 
+
+#[cfg(feature = "sgx")]
+use sgx_tstd as std;
+
+#[cfg(feature = "sgx")]
+use std::vec::Vec;
+
+/// payload of block that needs to be encrypted
+#[derive(PartialEq, Eq, Clone, Encode, Decode, Debug)]
+#[cfg_attr(feature = "std", derive(Serialize, Deserialize))]
+pub struct StatePayload {
+    state_hash_apriori: H256,
+    state_hash_aposteriori: H256,
+    /// encoded state update
+    state_update: Vec<u8>,
+}
+
+impl StatePayload {
+    /// get hash of state before block execution
+    pub fn state_hash_apriori(&self) -> H256 {
+        self.state_hash_apriori
+    }
+    /// get hash of state after block execution
+    pub fn state_hash_aposteriori(&self) -> H256 {
+        self.state_hash_aposteriori
+    }
+    /// get encoded state update reference
+    pub fn state_update(&self) -> &Vec<u8> {
+        &self.state_update
+    }
+    pub fn new(apriori: H256, aposteriori: H256, update: Vec<u8>) -> StatePayload {
+        StatePayload {
+            state_hash_apriori: apriori,
+            state_hash_aposteriori: aposteriori,
+            state_update: update,
+        }
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -300,5 +342,25 @@ mod tests {
         );
 
         assert!(signed_call.verify_signature(&mrenclave, &shard));
+    }
+
+    #[test]
+    fn new_payload_works() {
+        // given
+        let state_hash_apriori = H256::random();
+        let state_hash_aposteriori = H256::random();
+        let state_update: Vec<u8> = vec![];
+
+        // when
+        let payload = StatePayload::new(
+            state_hash_apriori.clone(),
+            state_hash_aposteriori.clone(),
+            state_update.clone(),
+        );
+
+        // then
+        assert_eq!(state_hash_apriori, payload.state_hash_apriori());
+        assert_eq!(state_hash_aposteriori, payload.state_hash_aposteriori());
+        assert_eq!(state_update, *payload.state_update());
     }
 }
